requirements.txt
# Core dependencies
requests>=2.32.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web scraping (dynamic content)
playwright>=1.40.0

# Document generation
markdown>=3.5.0

# Language detection
langdetect>=1.0.9

# Data validation
pydantic>=2.0.0

# UI (optional, for Streamlit interface)
streamlit>=1.30.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0

# Utilities
lxml>=4.9.0